waits for the timeout by default
This addresses a subtle bug in the test logic. In the initial implementation of
the first test, we're waiting for 1 notification, so we used the default
`num_notifications` value of 1. Because of the way NotificationWaiter
works, though, that means it will stop waiting once it observes a single
notification. That means that as long as the first notification meets
the expectations of the test, the test will pass, no matter how  many
more notifications would have been observed. The assertion "the
NotificationWaiter must have only observed one notification" really
means "the node received 1 *or more* notifications", since the
NotificationWaiter stops waiting after the first.

Similarly, the second test's initial implementation will succeed when
the first two notifications are 'DOWN' and 'UP', regardless of what
notifications follow.

To avoid this, this commit changes NotificationWaiter's default
behavior. Instead of stopping after a predefined number of commits, it
waits for the complete timeout. the `num_notifications` argument is now
optional.

It also removes the logic that raises an error if the waiter observes an
unexpected number of notifications. Those assertions should be made in
the tests.