Fix failing compressed commitlog size test
There are a few adjustments here.

First, the expected total commitlog size was changed from 23MB to 42MB.
42 seemed like a good choice for a few of reasons:
 * If we assume the compressed size is 0.7 times the original size
   which other parts of the test do, then 0.7 * 60 == 42
 * The small_compressed_segment_size_test also expects a size of 42MB,
   and we can assume the compression ratio is similar even with a
   smaller commitlog segment size
 * The actual size in practice is about 43MB

Second, the tolerated error of 5% is fairly low for compressed data.
I've conditionally increased the tolerated error when compression is in
use.  Theoretically, CASSANDRA-9095 should eliminate the need to do this
(at least for segment sizes).