Ignore jamm-related error log for offline test in 2.0
Since jamm is not in the classpath for the 2.0 setup during the offline
tools test, Cassandra will log the following error message:

"Unable to initialize MemoryMeter (jamm not specified as javaagent).
This means Cassandra will be unable to measure object sizes accurately
and may consequently OOM."

This message can be safely ignored.