improves comparison in compaction throughput
Comparison of strings representing numbers is too fragile. If we ever
change the threshold to '20', for instance, then we'll run into trouble:

    $ python
    >>> '20' >= '100'
    True

This patch changes the values to floats before comparison:

    $ python
    >>> float('20') >= float('100')
    False

This patch also allows more leeway in the comparison; compaction
throughput is allowed to be up to 125% of the specified throughput.

This patch also uses a regex rather than find to extract the compaction
throughput value from the log. After converting avgthroughput to a
float, the test started failing on output lines like

    INFO  [CompactionExecutor:1] 2015-06-16 14:56:30,555 CompactionTask.java:223 - Compacted (64f6aa00-1459-11e5-b4e8-bf89e1b7b083) 3 sstables to [/tmp/dtest-Pk59zB/test/node1/data/keyspace1/standard1-50ace0a0145911e5b4e8bf89e1b7b083/la-4-big,] to level=0.  32,073,902 bytes to 28,100,000 (~87% of original) in 2,205ms = 12.153401MB/s.  0 total partitions merged to 100,000. Partition merge counts were {1:85858, 2:14142, }

Note the earlier-than-expected equals sign. This regex captures the actual value more carefully.