disambiguates name for hinted handoff test
This doesn't test hinted handoff only, it depends on and tests new
configuration features of hinted handoff. This caused a little confusion
on IRC, so I've made the name of the test class indicate the feature it
tests.

Looking through the test's history, I also noticed it was associated
with a Jira ticket, so I added that metadata as well.