fixes occasional failures in simple_rebuild_test
This removes the test's dependence on the outcome of a race.
Occasionally, the call to `nodetool rebuild` in the main thread would
start before the one in the secondary thread, making the call in the
secondary thread fail instead of the one in the main thread. The failure
is correct -- a `nodetool rebuild` call should fail if another is
running. However, the test only expected failure in the main thread.

Eliminating the race condition itself is...  probably impossible? Or at
least difficult. So, instead, the new logic checks that exactly one of
the calls throws the error, rather than requiring one or the other to.