Fix commitlog_test:TestCommitLog.test_commitlog_replay_on_startup
The test was grepping "0 replayed mutations" and checking we had no hit
to verify that some mutations were replayed. But that was actually
matching "20 replayed mutations" which wasn't the intent.
Added a space before the 0 since the message contains one and it avoids
the ambiguity, but I'm not even sure grepping the log is a good idea
in the first place.