catch OSError when shelling out to git
Previously this code assumed that `git` would be available in the
environment running `nosetests`. This is not always true; this was
brought to my attention by a user in a desktop Windows environment who
uses a `git` shell for some work but `cmd.exe` for `nosetests`. This PR
changes the function that shells out to `git` so it returns `None` if
you can't shell out.