s/xrange/range/g in cql_tests.py
`xrange` won't be around in Python 3. The benefit of using `xrange` is
that it just iterates over the values rather than constructing a list.
These are small enough lists that there really isn't a big worry about
that, so I figured I'd sed these out while there's churn on this file.