remove upgrade_through dep on schema_metadata_test
this has been the source of a number of problems and should be
removed for simplification.

each of these test modules is built with a different purpose in mind
and having them tangled has become an issue.

this needs to be followed up with proper expansion of schema coverage
within upgrade_through_versions or in another test built for the express
purpose of upgrade testing