Fix test for CASSANDRA-10902
It seems CASSANDRA-10902 was unintentionally dropped from trunk before
3.3 was cut, but the test for it in configuration_test has not been
failing. The ccm change related to CASSANDRA-6696 was masking this as
the test assumed the node's (only) data directory would be named 'data'.
As this is no longer true, the cache location was not a sub directory
of a data directory and the error condition the test is checking for
was never triggered.