change assertion logic for ">0 mutations replayed"
The logic needed to change -- the existing logic asserted that " 0
replayed mutations" was not in C*'s logs, which is not the same as
asserting that some mutations were replayed. This improves the logic so
that it a) asserts there was at least one "replayed mutations" line
logged, and b) asserts at least one of those indicated there were a
non-zero number of mutations replayed.

This introduces a dependency on the `parse` library, which seems very
nice.