use loop entry condition instead of `break`
Before, the loop exited with a `break` if the replay finished. Now, that
logic is explicitly added as part of the `while` condition.

Before, if the loop exited because the timer assertion failed, the
`else:` block would execute and the test would fail. Now, the loop can
exit either because replay has finished or because the timer assertion
failed, so the `fail` call only happens if replay has not finished.