simplify loop exit/timeout assertion logic
This further simplifies the while loop logic. Before, the loop could
exit if replay finished or if the operation "timed out". Then, if replay
didn't finish (implying the operation timed out), the test fails. Now,
if the "timeout" occurs, the assertion will fail and it will exit the
loop by failing the test.