CASSANDRA-16433: Cumulative fix for upgrade tests
[070901c7572afee1cf649501e2161f3bfdb3e950] CASSANDRA-16433: Fix Cassandra version families

[52ca8cbb7c1f6dd4c34e70d5ab0b821abb5ed393] CASSANDRA-16433: add some changes

[7b022665f1b4f42961c455ac71602456cc07e760] CASSANDRA-16433: Upgrade manifest update
1. There is no reason to test upgrades from unsupported indev versions

2. There is no reason to test upgrade to current if indev is still supported

[11b7e30e33a7bcfaa7072909827d5eeb2314d3a5] CASSANDRA-16433: Removed OVERRIDE_MANIFEST
It seems to be unused. If you need to change
something for your own you need to modify
upgrade_manifest.py anyway, so why not modify
the current manifest?

[b8d8db5f525d7f37aafd573cacfee702ec50a5ff] CASSANDRA-16433: Fix the problems caused by C-15193
In particular, CASSANDRA-15193 enforced protocol v3 during upgrade from pre 3.0 versions so we need to stick with 3 for upgrades from Cassandra 2.2. Since we do not support 2.2, we can just generalize and use v3 for Cassandra 2.2 in all cases.

Another approach would be to leave the protocol unset. It would automatically negotiate the right protocol but at the same time it would generate warnings from the driver

[b08274c7b2a4f44759d619ace2ec245008ff9aff] CASSANDRA-16433: Fix broken thrift_upgrade_test.py

[f140382bd1479a6673da32dafcb7e724568a3d94] CASSANDRA-16433: Update manifest

[3ea69899109057360426d342e09271e112d8aac3] CASSANDRA-16433: Fix hardcoded references to gitbox

[c69ce38f5ac551a8b653d6026d98dcc3bf28bd66] CASSANDRA-16433: Remove unnecessary implicit upgrade scenarios
When we test upgrades with --upgrade-target-version-only we expect to test upgrades from the target version and to the target version. This obviously means we want test those upgrade paths where the target (indev) version was the origin or destination version.

The problem which this commit fixes is that the filter was applied only for the version family but not on the variant. Thus, say testing some development branch of 3.11 would include running completely unrelated upgrade scenarios current-3.0 -> *-trunk.

[6d2f183b09551d211515b665f17843428093e5ab] CASSANDRA-16433: fix upgrade_udtfix_test.py
Those tests should be run since 3.11.6 (see CASSANDRA-15035)

[3799538b036a173cfb128e327d646f874fe6a3e1] CASSANDRA-16433: fix a couple of mistakes

[6f1a0f1d6d9d3bdad7c3ad0ec662b45942eb36bc] CASSANDRA-16433: try to fix the flakiness in upgrade_through_versions

[028fab3c93dcc3becc19c1951e1b052885a51b5b] Apply Tomek's suggestion