Fix unexpectedly ignored dtests
This commit fixes few unexpected behaviors, which were:
1. If any test class is marked with `resource_intensive`, then all tests
   in the file were treated as resource intensive.
2. If one test class is marked with `vnodes` and another test class is
   marked with `no_vnodes` in the same file, then all tests were
   skipped in the file.

The unexpected behaviors are due to marking all tests in a file if any
class is assigned the mark. This commit fixes that marks on other
classes in the file do not affect a test, which is not marked and
neither its class. This is also applied to upgrade tests, so the code
is simplified and behavior is unified.

Since the upgrade tests were utilizing that the marking one class is
propagated to all classes in the file, they were revisited. In many
cases the upgrade tests were already included, because their classes
inherited from a class, which already was marked. In such cases
unnecessary marks were removed.

patch by Ruslan Fomkin; reviewed by Andrés de la Peña and Ekaterina
Dimitrova for CASSANDRA-16841